requirements_test.txt
pip==10.0.1
bumpversion==0.5.3
wheel==0.31.1
watchdog==0.8.3
flake8==3.5.0
tox==3.0.0
coverage==4.5.1
Sphinx==1.7.5
twine==1.11.0


pytest==3.4.2
pytest-runner==2.11.1
pytest-mock==1.10.0
pytest-profiling==1.3.0
requests-mock==1.5.0
